re #418: Moved planner tests into query test
